requirements.txt
pandas
requests
python-dotenv
schedule